restructure of files
