--- mentorship/src/assets/components/Buttton/Button.css
@@ -1,9 +1,0 @@
-.btn-info {
-    color:white;
-    background-color:rgb(21, 129, 129);
-    padding:10px;
-    border-radius:10px ;
-    margin:5px;
-
-
-}
