--- mentorship/src/assets/components/Buttton/Button.jsx
@@ -1,12 +1,0 @@
-import React from 'react'
-import "./Button.css"
-
-function Button({text="click", btnClass, onClick}) {
-  return (
-    <button className={`btn ${btnClass}`} onClick={onClick}>
-      {text}
-    </button>
-  )
-}
-
-export default Button
