--- mentorship/src/assets/components/Navbar/navbar.css
@@ -1,116 +1,0 @@
-*{
-    margin:0;
-    padding: 0;
-    box-sizing: border-box;
-
-}
-
-.navbar{
-   
-    display: flex;
-    justify-content:flex-end;
-    align-items: center;
-    background-color:rgb(21, 129, 129);
-    height: 75px;
-    color:white;
-
-}
-
-.logo{
-    font-size: 20px;
-}
-
-.home,
-.about,
-.apply,
-.login,
-.contact{
-    text-decoration: none;
-    color: white;
-    font-size: 20px;
-    padding: 10px 20px;
-    margin: 0 10px;
-
-
-}
-
-.nav-links{
-    display: flex;
-    list-style: none;
-    justify-content: center;
-    width:70%
-   
-
-}
-
-.login{
-    border: 1px solid yellow;
-    border-radius: 5px;
-
-}
-
-.mobile-menu-icon{
-    display:none;
-}
-
-
-
-@media screen and (max-width:780px){
-
-    .logo{
-        display:flex;
-        position: absolute;
-        top: 15px;
-        left:35pX;
-
-    }
-
-    .nav-links{
-        display:none;
-      
-    }
-
-    .nav-links-mobile{
-        position: absolute;
-        display: block;
-        list-style: none;
-        background-color:#005469;
-        left: 0;
-        top: 75px;
-        transition: all 0.5s ease-out;
-        width: 100%;
-
-
-    }
-
-    .home,
-    .about,
-    .apply,
-    .login,
-    .contact{
-        color:white;
-        text-align: center;
-        padding: 32px;
-        width:100%;
-        transition: all 0.5s ease;
-    }
-
-
-    .login{
-        border: none;
-    }
-
-    .mobile-menu-icon{
-        display:block;
-        position: absolute;
-        font-size: 30px;
-        color:white;
-        background-color:#00828c;
-        border:none;
-        top:15px;
-        right:25px;
-    }
-
-
-
-}
