--- mentorship/src/assets/components/Pages/AboutPage.jsx
@@ -1,11 +1,0 @@
-const AboutPage = () =>{
-   
-    return(
-
-       <div>
-       <h1>About</h1>
-       </div>
-    )
-}
-
-export default AboutPage
