--- mentorship/src/assets/components/Pages/ApplyPage.jsx
@@ -1,13 +1,0 @@
-import ApplyForm from "../ApplyForm/ApplyForm"
-
-const ApplyPage = () =>{
-   
-    return(
-
-       <div>
-       <ApplyForm/>
-       </div>
-    )
-}
-
-export default ApplyPage
