--- mentorship/src/assets/components/Pages/ContactPage.jsx
@@ -1,11 +1,0 @@
-const ContactPage = () =>{
-   
-    return(
-
-       <div>
-       <h1>Contact</h1>
-       </div>
-    )
-}
-
-export default ContactPage
