--- mentorship/src/assets/components/Pages/Home.jsx
@@ -1,11 +1,0 @@
-const HomePage = () =>{
-   
-    return(
-
-       <div>
-       <h1>Home</h1>
-       </div>
-    )
-}
-
-export default HomePage
