--- mentorship/src/assets/components/Pages/LoginPage.jsx
@@ -1,13 +1,0 @@
-import LoginForm from "../LoginForm/LoginForm"
-
-const LoginPage = () =>{
-   
-    return(
-
-       <div>
-         <LoginForm/>
-       </div>
-    )
-}
-
-export default LoginPage
