+++ mentorship/src/components/LoginForm/LoginForm.jsx
@@ -1,4 +1,4 @@
-import Button from "../Buttton/Button";
+import Button from "../../assets/components/Buttton/Button";
 import "./loginForm.css";
