+++ mentorship/src/components/NavBar/NavBar.jsx
@@ -1,7 +1,7 @@
 import { Link, Outlet } from "react-router-dom";
 import { useEffect, useState } from "react";
 
-import useAuth from "../hooks/use-auth";
+import useAuth from "../../hooks/use-auth";
 
 import "./NavBar.css";
