+++ mentorship/src/main.jsx
@@ -2,13 +2,16 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 
-import HomePage from './assets/components/pages/Home';
 import { AuthProvider } from './components/AuthProvider.jsx';
-import AboutPage from './assets/components/pages/AboutPage';
-import ApplyPage from './assets/components/pages/ApplyPage';
-import LoginPage from './assets/components/pages/LoginPage';
-import ContactPage from './assets/components/pages/ContactPage';
-import NavBar from './assets/components/navbar/navbar';
+import NavBar from './components/NavBar/NavBar.jsx';
+import HomePage from './pages/HomePage/HomePage.jsx';
+
+
+
+// import AboutPage from './assets/components/pages/AboutPage';
+// import ApplyPage from './assets/components/pages/ApplyPage';
+// import LoginPage from './assets/components/pages/LoginPage';
+// import ContactPage from './assets/components/pages/ContactPage';
 
 const router = createBrowserRouter([
   {
@@ -16,10 +19,10 @@ const router = createBrowserRouter([
     element:<NavBar />,
     children: [
       {path:"/",element:<HomePage />},
-      {path:"/about",element:<AboutPage />},
-      {path:"/apply",element:<ApplyPage />},
-      {path:"/login",element:<LoginPage />},
-      {path:"/contact",element:<ContactPage />},
+      // {path:"/about",element:<AboutPage />},
+      // {path:"/apply",element:<ApplyPage />},
+      // {path:"/login",element:<LoginPage />},
+      // {path:"/contact",element:<ContactPage />},
     ],
   },
 ]);
